% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/main.R
\name{clustANDscore}
\alias{clustANDscore}
\title{Cluster and Score}
\usage{
clustANDscore(data, genotypes, ...)
}
\arguments{
\item{data}{}
}
\value{
to_r
}
\description{
The function \code{\link{cluster}}
}
\examples{
clustANDscore()
}

