% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/main.R
\name{clustqtl_cluster}
\alias{clustqtl_cluster}
\title{Cluster data}
\usage{
clustqtl_cluster(data, cluster_method = c("fuzzy")[1],
  distance = c("cosine", "euclidean")[1])
}
\arguments{
\item{data}{}
}
\value{
Clustering object of input data
}
\description{
The function \code{\link{cluster}}
}

