% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/main.R
\name{clustqtl_score}
\alias{clustqtl_score}
\title{Score clustering}
\usage{
clustqtl_score(clusters, genotype_vector, verbose = FALSE)
}
\arguments{
\item{data}{}
}
\value{
Score representing distance between the actual clustering
 and the one proposed by model using binomial distribution
}
\description{
The function \code{\link{cluster}}
}

