% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utilities.R
\name{combineMarkers}
\alias{combineMarkers}
\title{combineMarkers}
\usage{
combineMarkers(genotypes, markers, limit_strains = NULL,
  limit_markers = NULL, na.rm = TRUE, rm_type = c("marker", "strain")[1],
  collpase_markers = TRUE, marker_rename = FALSE, impute_markers = TRUE,
  clean_markers = TRUE)
}
\arguments{
\item{na.rm}{Remove rows that generate NaN values}

\item{x}{A matrix}
}
\value{
1-cosine similarity matrix
}
\description{
Combine markers if available genotypes at adjacent locations
are identical
}

