% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utilities.R
\name{cosine}
\alias{cosine}
\title{Cosine similarity}
\usage{
cosine(x, y)
}
\arguments{
\item{x}{A vector}

\item{y}{Another vector}
}
\value{
Cosine similarity
}
\description{
Calculate cosine similarity between two vectors, \code{x} and \code{y}
}
\examples{
x = rnorm(100,1)
y = rnorm(100,1)
cosine(x,y)
}

