% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utilities.R
\name{cosineDist}
\alias{cosineDist}
\title{cosineDist}
\usage{
cosineDist(x, na.rm = TRUE)
}
\arguments{
\item{x}{A matrix}

\item{na.rm}{Remove rows that generate NaN values}
}
\value{
1-cosine similarity matrix
}
\description{
Calculate the cosine similarity between all rows of data matrix
 \code{x}
}
\examples{
x = do.call(rbind,lapply(1:5,function(i){rnorm(5,1)}))
cosineDist(x)
}

