% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/preprocess.R
\name{find_sigCounts}
\alias{find_sigCounts}
\title{Find significant counts in data matrix by resampling}
\usage{
find_sigCounts(x, nresamples = 1000, alpha = 0.01)
}
\arguments{
\item{x}{An \code{m x n} matrix of peak intensities where the rows, \code{m},
contains individuals, (e.g., strains)
and the columns, \code{n}, contains observations (e.g., positions)}

\item{nresamples}{Number of resamples to compute.}

\item{alpha}{Significance level at which to reject the null hypothesis, Default = 0.01}
}
\value{
A count value at which the null hypothesis can be rejected given \code{alpha}.
}
\description{
The function \code{\link{find_sigCounts}} returns a count cutoff at a specified
 significance value, \code{alpha}, by resampling. The null model specifies a uniform random
 distribution of counts across the sampling interval. The returned value signifies
 the count value at which the null hypothesis can be rejected given some confidence
 value, \code{alpha}
}
\examples{
# S. cerevisiae chr01 3' isoform counts
x = subsetByOverlaps(tx3_counts_chr01,tx_3utr_annotations_yeast_chr01[1])
x = t(as.matrix(mcols(x)))
find_sigCounts(x)
}

