% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utilities.R
\name{findQTLPeaks}
\alias{findQTLPeaks}
\title{Find QTL peaks}
\usage{
findQTLPeaks(qtls, mrk, pcutoff = 0.05, peak_sigma = 25,
  peak_threshold = 1, ...)
}
\arguments{
\item{qtls}{}
}
\value{
Data Frame of peaks
}
\description{
The function \code{\link{findQTLPeaks}}
}

