% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot.R
\name{format4manhattan}
\alias{format4manhattan}
\title{Permute}
\usage{
format4manhattan(qtls, mrk, nresample = 10000)
}
\arguments{
\item{data}{}
}
\value{
Modified mrk
}
\description{
The function \code{\link{cluster}}
}
\examples{
format4manhattan()
}

