% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utilities.R
\name{getGenotypes}
\alias{getGenotypes}
\title{getGenotype}
\usage{
getGenotypes(marker, geno)
}
\arguments{
\item{data}{}
}
\value{
genotypes
}
\description{
The function \code{\link{cluster}}
}
\examples{
getGenotypes()
}

