% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/preprocess.R
\name{granges2matrix}
\alias{granges2matrix}
\title{Converts GRanges object to matrix with real coordinates}
\usage{
granges2matrix(x, annot = NULL)
}
\arguments{
\item{x}{An \code{m x n} matrix of peak intensities where the rows, \code{m},
contains individuals, (e.g., strains)
and the columns, \code{n}, contains observations (e.g., positions)}

\item{nresamples}{Number of resamples to compute.}

\item{alpha}{Significance level at which to reject the null hypothesis, Default = 0.01}
}
\value{
A count value at which the null hypothesis can be rejected given \code{alpha}.
}
\description{
Some more info
}
\examples{
# S. cerevisiae chr01 3' isoform counts
x = GenomicRanges::subsetByOverlaps(tx3_counts_chr01,tx_3utr_annotations_yeast_chr01[1])
x = t(as.matrix(as.data.frame(GenomicRanges::mcols(x))))
find_sigCounts(x)
}

