% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/main.R
\name{permute}
\alias{permute}
\title{Permute}
\usage{
permute(genotypes, clusters, nresample = 10000, ...)
}
\arguments{
\item{data}{}
}
\value{
permuted scores
}
\description{
The function \code{\link{cluster}}
}
\examples{
permute()
}

