% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot.R
\name{plotManhattan}
\alias{plotManhattan}
\title{Permute}
\usage{
plotManhattan(qtls, mrk, qtl_name = "", trx_annot = NULL, cutoff = 3,
  gene_annot_range = c(5000, 5000), ...)
}
\arguments{
\item{data}{}
}
\value{
plot
}
\description{
The function \code{\link{cluster}}
}
\examples{
plotManhattan()
}

