% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot.R
\name{plotPeakProfile}
\alias{plotPeakProfile}
\title{Plot peak profiles}
\usage{
plotPeakProfile(data, genotypes, marker, peak_sigma = 2, peak_threshold = 1)
}
\arguments{
\item{qtls}{A two column matrix with each row containing a pvalue for
every marker in \code{mrk}}
}
\value{
plot
}
\description{
The function \code{\link{cluster}}
}

